% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fomtdata}
\alias{fomtdata}
\title{Total CQA with ascorbic acid (FOMT model)}
\format{
A data frame with 8 rows and 2 columns:
\describe{
\item{time_h}{Time in hours}
\item{tCQA_AA}{Normalized concentration of total CQA measured at each
time point}
}
}
\source{
Yusaku N. and Kuniyo I. (2013)
\emph{Degradation Kinetics of Chlorogenic
Acid at Various pH Values and Effects of Ascorbic Acid and Epigallocatechin
Gallate on Its Stability under Alkaline Conditions},
Journal of Agricultural and Food Chemistry, \doi{10.1021/jf304105w},
fig 2D, solid diamonds
}
\usage{
fomtdata
}
\description{
Degradation data of 1.2 mM 5-caffeoylquinic acid (5-CQA) in the presence of
1.2 mM of ascorbic acid at 37°C. The data refer to total CQA concentration.
}
\keyword{datasets}
