% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ord1}
\alias{ord1}
\title{First order kinetic data}
\format{
A data frame with 6 rows and 3 columns:
\describe{
\item{t}{time}
\item{concentration}{simulated concentration data at each time point}
\item{std.error}{simulated experimental error}
}
}
\usage{
ord1
}
\description{
Synthetic data from a first-order kinetic model with k=0.7
}
\keyword{datasets}
