% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urfa}
\alias{urfa}
\title{Urfa pepper ascorbic acid degradation (2-nd order)}
\format{
A data frame with 8 rows and 4 columns:
\describe{
\item{time_min}{time in minutes}
\item{AA_55}{normalized concentration of ascorbic acid of Urfa peppers
dehydrated at 55°C}
\item{AA_65}{normalized concentration of ascorbic acid of Urfa peppers
dehydrated at 65°C}
\item{AA_75}{normalized concentration of ascorbic acid of Urfa peppers
dehydrated at 75°C}}
}
\source{
Ş. Dağhan, A. Yildirim, F. Mehmet Yilmaz, H. Vardin and
M. Karaaslan (2018)
\emph{The effect of temperature and method of drying on isot (Urfa pepper)
and its Vitamin C degradation kinetics},
Italian Journal of Food Science, DOI:
\href{http://dx.doi.org/10.14674/IJFS-1070}{10.14674/IJFS-1070},
fig 5, hot-air
}
\usage{
urfa
}
\description{
Data describing the degradation kinetics of ascorbic acid during dehydration
of Urfa peppers. The peppers were treated with hot air at 55, 65 and 75 °C.
}
\keyword{datasets}
